+++ test/test_tcp_server/test_connectivity.py
@@ -40,3 +40,20 @@ def test_tcp_server_lock():
     server.server_close()
     server.shutdown()
     t.join()
+
+
+def test_tcp_server_connection_reset_error():
+    server_address = ("127.0.0.1", 19000)
+    server = TcpFakeServer(server_address, server_type="redis")
+    t = Thread(target=server.serve_forever, daemon=True)
+    t.start()
+    time.sleep(0.1)
+
+    with redis.Redis(*server_address) as r:
+        r.rpush("test", b"foo")
+
+    with redis.Redis(*server_address) as r:
+        assert r.rpop("test") == b"foo"
+
+    server.shutdown()
+    t.join()